+++ package.json
@@ -22,6 +22,7 @@
     "react-native-screens": "^3.3.0",
     "react-redux": "^7.2.4",
     "redux": "^4.1.0",
+    "redux-form": "^8.3.7",
     "redux-recompose": "^3.0.0",
     "redux-thunk": "^2.3.0"
   },
