+++ src/redux/index.js
@@ -1,4 +1,5 @@
 import {combineReducers} from 'redux';
+import {reducer as formReducer} from 'redux-form';
 
 import {historyReducer} from './history/historyReducer';
 import {userReducer} from './auth/userReducer';
@@ -8,4 +9,5 @@ export default combineReducers({
   history: historyReducer,
   user: userReducer,
   historyApi: historyApiReducer,
+  form: formReducer,
 });
